+++ modules/luabundle.lua
@@ -1,13 +1,17 @@
 -------------------------------------------------------------------------------
--- Bundle module
+-- Lua bundler module
 -- Sledmine
--- Bundler for lua mod projects
+-- Bundle features for modular lua projects
 -------------------------------------------------------------------------------
 local glue = require "glue"
 local json = require "cjson"
+local pjson = require "pretty.json"
 local bundle = require "bundle"
+local fs = require "fs"
 
 local codeBundler = require "lib.codeBundler"
+
+local luabundler = {}
 
 ---@class bundle
 ---@field name string
@@ -16,7 +20,8 @@ local codeBundler = require "lib.codeBundler"
 ---@field main string
 ---@field output string
 
-local function bundler(bundleName, compile)
+--- Bundle a modular lua project using luacc implementation
+function luabundler.bundle(bundleName, compile)
     --  TODO Add compilation feature based on the lua target
     if (bundleName) then
         bundleName = bundleName .. "Bundle.json"
@@ -37,13 +42,15 @@ local function bundler(bundleName, compile)
                 cprint("done.")
                 if (compile) then
                     cprint("Compiling project... ", true)
-                    local compile = project.target:gsub("lua", "luac") .. " -o " .. project.output
+                    local compile = project.target:gsub("lua", "luac") .. " -o " ..
+                                        project.output
                     compile = compile .. " " .. project.output
                     local compileResult = os.execute(compile)
                     if (compileResult) then
                         cprint("done.")
                     else
-                        cprint("Error, compilation process encountered one or more errors!")
+                        cprint(
+                            "Error, compilation process encountered one or more errors!")
                     end
                 end
             else
@@ -53,8 +60,27 @@ local function bundler(bundleName, compile)
         end
         return true
     end
-    cprint("Warning, There is not a " .. bundleFileName .. " in this folder, be sure to create one.")
+    cprint("Warning, there is not a " .. bundleFileName ..
+               " in this folder, be sure to create one.")
     return false
 end
 
-return bundler
+--- Attempt to create a manifest file template
+function luabundler.template()
+    if (not fs.is("manifest.json")) then
+        local template = {
+            name = "",
+            target = "lua53",
+            include = {"modules\\"},
+            modules = {""},
+            main = "",
+            output = "dist\\.lua"
+        }
+        glue.writefile("manifest.json", pjson.stringify(template, nil, 4), "t")
+        cprint("Success, manifest.json template has been created successfully.")
+    else
+        cprint("Warning, there is already a manifest in this folder!")
+    end
+end
+
+return luabundler